refactor: api endpoint 개선
