+++ src/main/resources/static/js/ranking.js
@@ -1,5 +1,5 @@
 document.addEventListener('DOMContentLoaded', () => {
-    requestRead('/theme-ranking') // 인기 테마 목록 조회 API endpoint
+    requestRead('/theme/ranking') // 인기 테마 목록 조회 API endpoint
         .then(render)
         .catch(error => console.error('Error fetching times:', error));
 });
